Reverted change to moving ConfigActionsTest
