+++ src/test/java/com/blackducksoftware/integration/hub/alert/web/actions/distribution/EmailConfigActionsTest.java
@@ -24,7 +24,6 @@ import com.blackducksoftware.integration.hub.alert.datasource.relation.repositor
 import com.blackducksoftware.integration.hub.alert.mock.entity.MockEmailEntity;
 import com.blackducksoftware.integration.hub.alert.mock.model.MockEmailRestModel;
 import com.blackducksoftware.integration.hub.alert.web.ObjectTransformer;
-import com.blackducksoftware.integration.hub.alert.web.actions.ActionsTest;
 import com.blackducksoftware.integration.hub.alert.web.actions.ConfiguredProjectsActions;
 import com.blackducksoftware.integration.hub.alert.web.actions.NotificationTypesActions;
 import com.blackducksoftware.integration.hub.alert.web.model.distribution.EmailGroupDistributionRestModel;
